Figuring out problem with delayed delete of room
The problem is, that when user is the only player in the game, the
transfer of ownership is not possible. So the game room has to be
destroyed on disconnect event, which can also happen on page refresh.
This means that the GET for index will be triggered, which will in turn
load the current game rooms. The problem is that this happens before the
disconnect event! I can see only two solutions. The first is delayed
delete, where other players can in the mean time take ownership of the
game room with joining the game. The second is better. Create a session
(cookie) which will store game rooms for the client. On reconnect take
back the ownership of the room, but before that check if somebody else
got it there before. The timeout before destroying the game room should
be a few seconds.