Implemented resize of game display
Implemented logic to handle resizing of game display according to user
screen size. The logic is the following. It gets the min of the width of
the column, that holds the game display vs. the user screen height. Then
it computes the maximum number of columns or rows of stones. With this
information we can then compute the multiplier which will in turn be
used to compute the height and width of the element.